refactor(api): move codes package
