+++ api/internal/store/entity/shipping.go
@@ -5,7 +5,7 @@ import (
 	"errors"
 	"time"
 
-	"github.com/and-period/furumaru/api/internal/store/codes"
+	"github.com/and-period/furumaru/api/internal/codes"
 	set "github.com/and-period/furumaru/api/pkg/set/v2"
 	"github.com/and-period/furumaru/api/pkg/uuid"
 	"gorm.io/datatypes"
